refactor(utils): split utilities into package
